pyproject.toml
[tool.poetry]
name = "cdk-lab-s3-batch"
version = "0.1.0"
description = ""
authors = ["Violet Vivirand <purpleslovemail@gmail.com>"]
readme = "README.md"
packages = [{include = "cdk_lab_s3_batch"}]

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = ">=2.106.0"
constructs = ">=10.0.0,<11.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "6.2.5"
boto3 = "^1.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
